+++ packages/blob/src/index.js
@@ -70,3 +70,38 @@ export function isBlobURL( url ) {
 	}
 	return url.indexOf( 'blob:' ) === 0;
 }
+
+/**
+ * Downloads a file, e.g., a text or readable stream, in the browser.
+ * Appropriate for downloading smaller file sizes, e.g., < 5 MB.
+ *
+ * Example usage:
+ * 	const fileContent = JSON.stringify(
+ * 		{
+ * 			"title": "My Post",
+ * 		},
+ * 		null,
+ * 		2
+ * 	);
+ * 	const fileName = 'file.json';
+ *
+ * 	downloadBlob( 'file.json', fileContent, 'application/json' );
+ *
+ * @param {string} fileName    File Name.
+ * @param {string} content     File Content.
+ * @param {string} contentType File mime type.
+ * @return {HTMLElement}       The created anchor element.
+ */
+export function downloadBlob( fileName = 'wp-download', content, contentType ) {
+	const file = new window.Blob( [ content ], { type: contentType } );
+	const url = window.URL.createObjectURL( file );
+	const a = document.createElement( 'a' );
+	a.href = url;
+	a.download = fileName;
+	a.style.display = 'none';
+	document.body.appendChild( a );
+	a.click();
+	document.body.removeChild( a );
+	window.URL.revokeObjectURL( url );
+	return a;
+}