+++ packages/blob/src/test/index.js
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import { isBlobURL, getBlobTypeByURL } from '../';
+import { isBlobURL, getBlobTypeByURL, downloadBlob } from '../';
 
 describe( 'isBlobURL', () => {
 	it( 'returns true if the url starts with "blob:"', () => {
@@ -24,5 +24,33 @@ describe( 'getBlobTypeByURL', () => {
 
 	it( 'returns undefined if the url is not defined', () => {
 		expect( getBlobTypeByURL() ).toBe( undefined );
+	} );
+} );
+
+describe( 'downloadBlob', () => {
+	const originalURL = window.URL;
+	const createObjectURL = jest.fn().mockReturnValue( 'blob:pannacotta' );
+	const revokeObjectURL = jest.fn().mockReturnValue( false );
+	beforeEach( () => {
+		window.URL = {
+			createObjectURL,
+			revokeObjectURL,
+		};
+	} );
+
+	afterAll( () => {
+		window.URL = originalURL;
+	} );
+	it( 'returns expected HTML element without filename', () => {
+		downloadBlob( undefined, '{}', 'application/json' );
+		expect( createObjectURL ).toHaveBeenCalledWith( new window.Blob() );
+		expect( revokeObjectURL ).toHaveBeenCalled();
+	} );
+	it( 'returns expected HTML element with filename', () => {
+		expect(
+			downloadBlob( 'file.json', '{}', 'application/json' ).outerHTML
+		).toBe(
+			'<a href="blob:pannacotta" download="file.json" style="display: none;"></a>'
+		);
 	} );
 } );