+++ packages/edit-site/src/components/page-patterns/grid-item.js
@@ -36,6 +36,7 @@ import {
 } from '@wordpress/icons';
 import { store as noticesStore } from '@wordpress/notices';
 import { store as reusableBlocksStore } from '@wordpress/reusable-blocks';
+import { downloadBlob } from '@wordpress/blob';
 
 /**
  * Internal dependencies
@@ -50,25 +51,6 @@ import {
 import { store as editSiteStore } from '../../store';
 import { useLink } from '../routes/link';
 import { unlock } from '../../lock-unlock';
-
-/**
- * Downloads a file.
- * Also used in packages/list-reusable-blocks/src/utils/file.js.
- *
- * @param {string} fileName    File Name.
- * @param {string} content     File Content.
- * @param {string} contentType File mime type.
- */
-function download( fileName, content, contentType ) {
-	const file = new window.Blob( [ content ], { type: contentType } );
-	const a = document.createElement( 'a' );
-	a.href = URL.createObjectURL( file );
-	a.download = fileName;
-	a.style.display = 'none';
-	document.body.appendChild( a );
-	a.click();
-	document.body.removeChild( a );
-}
 
 const { useGlobalStyle } = unlock( blockEditorPrivateApis );
 
@@ -136,7 +118,7 @@ function GridItem( { categoryId, item, ...props } ) {
 			syncStatus: item.patternBlock.wp_pattern_sync_status,
 		};
 
-		return download(
+		return downloadBlob(
 			`${ kebabCase( item.title || item.name ) }.json`,
 			JSON.stringify( json, null, 2 ),
 			'application/json'